pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "slicks"
version = "0.2.3"
description = "The home baked data pipeline for Western Formula Racing"
readme = "README.md"
authors = [
  { name = "Western Formula Racing", email = "fsae.uwo@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "SQLAlchemy>=2.0.0",
    "cantools>=39.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.0.0",
    "tqdm>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "interrogate>=1.5.0",
]

[project.urls]
"Homepage" = "https://github.com/Western-Formula-Racing/wfr-telemetry"
"Documentation" = "https://western-formula-racing.github.io/wfr-telemetry/"
"Bug Tracker" = "https://github.com/Western-Formula-Racing/wfr-telemetry/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=slicks --cov-report=xml --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src/slicks"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if __name__ == .__main__.:"]
fail_under = 20

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
fail-under = 70
exclude = ["tests", "setup.py"]
verbose = 1
color = true
generate-badge = "."
badge-format = "svg"
